[EC 1.4.x] Forward-compatibility for andra.ci.opscode.us
In order to build EC 1.4.x andra, this commit adds forward support
to the omnibus cli command to mimic the differences in the new cli.
While in compatibility mode, the cli will also rewrite the newer-style
configuration files to conform to the old block-style config that
this version of omnibus can understand.